refactor: directory pattern at blog apps
