+++ apps/blog/src/components/DateAndCategoryLink/index.tsx
@@ -1,26 +1,1 @@
-import NextLink from 'next/link';
-import { Link } from '@nextui-org/react';
-
-interface Props {
-  date: string;
-  category?: string;
-}
-
-function DateAndCategoryLink({ date, category }: Props) {
-  return (
-    <>
-      {date}
-      {category && (
-        <>
-          {' '}
-          at{' '}
-          <NextLink href={`/category/${category}`} passHref>
-            <Link color="primary">{category}</Link>
-          </NextLink>{' '}
-          category
-        </>
-      )}
-    </>
-  );
-}
-export default DateAndCategoryLink;
+export { default } from './DateAndCategoryLink';
